moved RawOrder to tof package
